🐛 fix(history): build 에러 수정
