+++ src/app/history/[id]/page.tsx
@@ -9,6 +9,7 @@ interface OrderItem {
   name: string;
   quantity: number;
   price: number;
+  imageUrl: string;
 }
 
 interface OrderDetail {
@@ -100,7 +101,7 @@ const OrderDetailPage = () => {
               >
                 <div className='flex items-center gap-4'>
                   <img
-                    src='/images/coke-zero.png'
+                    src={item.imageUrl}
                     alt={item.name}
                     className='w-14 h-14 rounded-md'
                   />
